requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-decouple==3.8
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.1
django-cors-headers==4.3.1